Feedback fixes
