+++ packages/core/admin/admin/src/components/Theme.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 
 import { DesignSystemProvider } from '@strapi/design-system';
 import { useIntl } from 'react-intl';
-import { ThemeName } from 'src/contexts/themeToggle';
 import { createGlobalStyle } from 'styled-components';
 
 import { useThemeToggle } from '../hooks/useThemeToggle';
@@ -11,27 +10,9 @@ interface ThemeProps {
   children: React.ReactNode;
 }
 
-type NonSystemThemeName = Exclude<ThemeName, 'system'>;
-
 const Theme = ({ children }: ThemeProps) => {
-  const { currentTheme, themes } = useThemeToggle();
+  const { currentTheme, themes, systemTheme } = useThemeToggle();
   const { locale } = useIntl();
-  const [systemTheme, setSystemTheme] = React.useState<NonSystemThemeName>();
-
-  // Listen to changes in the system theme
-  React.useEffect(() => {
-    const themeWatcher = window.matchMedia('(prefers-color-scheme: dark)');
-    setSystemTheme(themeWatcher.matches ? 'dark' : 'light');
-
-    themeWatcher.addEventListener('change', (event) => {
-      setSystemTheme(event.matches ? 'dark' : 'light');
-    });
-
-    // Cleanup on unmount
-    return () => {
-      themeWatcher.removeEventListener('change', () => {});
-    };
-  }, []);
 
   const computedThemeName = currentTheme === 'system' ? systemTheme : currentTheme;
