+++ packages/core/admin/admin/src/components/ThemeToggleProvider.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { DefaultTheme } from 'styled-components';
 
-import { ThemeToggleContext, ThemeName } from '../contexts/themeToggle';
+import { ThemeToggleContext, ThemeName, NonSystemThemeName } from '../contexts/themeToggle';
 
 const THEME_KEY = 'STRAPI_THEME';
 
@@ -22,6 +22,7 @@ interface ThemeToggleProviderProps {
 
 const ThemeToggleProvider = ({ children, themes }: ThemeToggleProviderProps) => {
   const [currentTheme, setCurrentTheme] = React.useState<ThemeName>(getDefaultTheme());
+  const [systemTheme, setSystemTheme] = React.useState<NonSystemThemeName>();
 
   const handleChangeTheme = React.useCallback(
     (nextTheme: ThemeName) => {
@@ -31,13 +32,30 @@ const ThemeToggleProvider = ({ children, themes }: ThemeToggleProviderProps) => 
     [setCurrentTheme]
   );
 
+  // Listen to changes in the system theme
+  React.useEffect(() => {
+    const themeWatcher = window.matchMedia('(prefers-color-scheme: dark)');
+    setSystemTheme(themeWatcher.matches ? 'dark' : 'light');
+
+    const listener = (event: MediaQueryListEvent) => {
+      setSystemTheme(event.matches ? 'dark' : 'light');
+    };
+    themeWatcher.addEventListener('change', listener);
+
+    // Remove listener on cleanup
+    return () => {
+      themeWatcher.removeEventListener('change', listener);
+    };
+  }, []);
+
   const themeValues = React.useMemo(() => {
     return {
       currentTheme,
       onChangeTheme: handleChangeTheme,
       themes,
+      systemTheme,
     };
-  }, [currentTheme, handleChangeTheme, themes]);
+  }, [currentTheme, handleChangeTheme, themes, systemTheme]);
 
   return <ThemeToggleContext.Provider value={themeValues}>{children}</ThemeToggleContext.Provider>;
 };
